Fix uploading asset when forced Id is present.

+++ lib/src/port/model/fields/asset_port_field.dart
@@ -33,7 +33,7 @@ class AssetPortField extends PortField<String?> {
       value = id;
     }
 
-    if (initialValue != null) {
+    if (initialValue != null && forcedAssetId == null) {
       await locate<AssetModule>().deleteAsset(initialValue!);
     }
